Updated structure for better downstream support
