+++ dio/CMakeLists.txt
@@ -1,3 +1,5 @@
+include(GNUInstallDirs)
+
 if (CMAKE_VERSION VERSION_LESS "3.0")
     project(rsdio)
 else()
@@ -6,27 +8,41 @@ endif()
 
 set(PROJECT_VERSION 3.0.4)
 
-include(GNUInstallDirs)
-include(GenerateExportHeader)
 option(BUILD_SHARED_LIBS "build as shared library" ON)
 
-set(diosrc rsdioimpl.cpp controllers/ite8783.cpp controllers/ite8786.cpp ../utils/tinyxml2.cpp)
-
-set(CMAKE_POSITION_INDEPENDENT_CODE ON)
+set(includedir ${PROJECT_SOURCE_DIR}/include)
+set(srcdir ${PROJECT_SOURCE_DIR}/src)
+set(src 
+    ${srcdir}/rsdioimpl.cpp 
+    ${srcdir}/controllers/ite8783.cpp 
+    ${srcdir}/controllers/ite8786.cpp 
+    ${rssdk_SOURCE_DIR}/utils/tinyxml2.cpp
+)
 
 set(CMAKE_CXX_VISIBILITY_PRESET hidden)
 set(CMAKE_VISIBILITY_INLINES_HIDDEN1)
 
 set(CMAKE_DEBUG_POSTFIX "d")
 
-add_library(rsdio ${diosrc})
+add_library(rsdio ${src})
 
-set_target_properties(rsdio PROPERTIES COMPILE_DEFINITIONS "RSDIO_VERSION_STRING=\"${PROJECT_VERSION}\"")
-set_target_properties(rsdio PROPERTIES VERSION ${PROJECT_VERSION})
+set_target_properties(rsdio PROPERTIES 
+    COMPILE_DEFINITIONS "RSDIO_VERSION_STRING=\"${PROJECT_VERSION}\""
+    VERSION ${PROJECT_VERSION}
+    POSITION_INDEPENDENT_CODE ON
+)
 
-target_include_directories(rsdio PRIVATE ${CMAKE_CURRENT_BINARY_DIR})
+include(GenerateExportHeader)
+generate_export_header(rsdio EXPORT_FILE_NAME ${CMAKE_CURRENT_BINARY_DIR}/exports/rsdio_export.h)
 
-generate_export_header(rsdio)
+target_include_directories(rsdio PRIVATE
+    ${rssdk_SOURCE_DIR}/utils
+)
+
+target_include_directories(rsdio PUBLIC
+    ${includedir}
+    ${CMAKE_CURRENT_BINARY_DIR}/exports
+)
 
 if(MSVC)
     target_compile_definitions(rsdio PUBLIC -D_CRT_SECURE_NO_WARNINGS)
