+++ poe/CMakeLists.txt
@@ -1,3 +1,5 @@
+include(GNUInstallDirs)
+
 if (CMAKE_VERSION VERSION_LESS "3.0")
     project(rspoe)
 else()
@@ -6,27 +8,43 @@ endif()
 
 set(PROJECT_VERSION 3.0.5)
 
-include(GNUInstallDirs)
-include(GenerateExportHeader)
 option(BUILD_SHARED_LIBS "build as shared library" ON)
 
-set(poesrc rspoeimpl.cpp controllers/pd69104.cpp controllers/pd69200.cpp controllers/ltc4266.cpp ../utils/tinyxml2.cpp ../utils/i801_smbus.cpp)
-
-set(CMAKE_POSITION_INDEPENDENT_CODE ON)
+set(includedir ${PROJECT_SOURCE_DIR}/include)
+set(srcdir ${PROJECT_SOURCE_DIR}/src)
+set(src 
+    ${srcdir}/rspoeimpl.cpp 
+    ${srcdir}/controllers/pd69104.cpp 
+    ${srcdir}/controllers/pd69200.cpp 
+    ${srcdir}/controllers/ltc4266.cpp 
+    ${rssdk_SOURCE_DIR}/utils/tinyxml2.cpp 
+    ${rssdk_SOURCE_DIR}/utils/i801_smbus.cpp
+)
 
 set(CMAKE_CXX_VISIBILITY_PRESET hidden)
 set(CMAKE_VISIBILITY_INLINES_HIDDEN1)
 
 set(CMAKE_DEBUG_POSTFIX "d")
 
-add_library(rspoe ${poesrc})
+add_library(rspoe ${src})
 
-set_target_properties(rspoe PROPERTIES COMPILE_DEFINITIONS "RSPOE_VERSION_STRING=\"${PROJECT_VERSION}\"")
-set_target_properties(rspoe PROPERTIES  VERSION ${PROJECT_VERSION})
+set_target_properties(rspoe PROPERTIES 
+    COMPILE_DEFINITIONS "RSPOE_VERSION_STRING=\"${PROJECT_VERSION}\""
+    VERSION ${PROJECT_VERSION}
+    POSITION_INDEPENDENT_CODE ON
+)
 
-target_include_directories(rspoe PRIVATE ${CMAKE_CURRENT_BINARY_DIR})
+include(GenerateExportHeader)
+generate_export_header(rspoe EXPORT_FILE_NAME ${CMAKE_CURRENT_BINARY_DIR}/exports/rspoe_export.h)
 
-generate_export_header(rspoe)
+target_include_directories(rspoe PRIVATE
+    ${rssdk_SOURCE_DIR}/utils
+)
+
+target_include_directories(rspoe PUBLIC
+    ${includedir}
+    ${CMAKE_CURRENT_BINARY_DIR}/exports
+)
 
 if(MSVC)
     target_compile_definitions(rspoe PUBLIC -D_CRT_SECURE_NO_WARNINGS)
