+++ poe/src/rspoeimpl.h
